64-bit transition part 7: libcmd builtins (#957)
This is the seventh of the thickfold patch series, which enables
ksh93 to operate within a 64-bit address space.

The parts of ksh93 affected by this commit are:
- The entirety of the libcmd builtins.

Remarks:
- Like the previous problematic getfsstat, sethostname also has a
  platform-dependent second argument that can be either signed or
  unsigned. This commit leaves the usage of that function (which
  passes a size_t value) as is.
- Moved the include directives in chmod and date to silence
  overlength string warnings.
- Removed unused macros from the fds code.
- Marked unused function parameters in grep to silence additional
  warnings.
- Use POSIX tcflag_t, speed_t and cc_t to silence warnings in stty.

Change in the number of warnings on Linux when compiling with clang
using -Wsign-compare -Wshorten-64-to-32 -Wsign-conversion
-Wimplicit-int-conversion: 1,578 => 1,189 => 37 (progression from
part 6 => part 7 => part 13)

Progresses https://github.com/ksh93/ksh/issues/592

+++ src/lib/libcmd/basename.c
@@ -2,7 +2,7 @@
 *                                                                      *
 *               This software is part of the ast package               *
 *          Copyright (c) 1992-2012 AT&T Intellectual Property          *
-*          Copyright (c) 2020-2024 Contributors to ksh 93u+m           *
+*          Copyright (c) 2020-2026 Contributors to ksh 93u+m           *
 *                      and is licensed under the                       *
 *                 Eclipse Public License, Version 2.0                  *
 *                                                                      *
@@ -69,7 +69,7 @@ static const char usage[] =
 static void namebase(Sfio_t *outfile, char *pathname, char *suffix, char termch)
 {
 	char *first, *last;
-	int n=0;
+	size_t n=0;
 	/* go to end of path */
 	for(first=last=pathname; *last; last++);
 	/* back over trailing '/' */
@@ -91,14 +91,14 @@ static void namebase(Sfio_t *outfile, char *pathname, char *suffix, char termch)
 		if(*first=='/')
 			first++;
 		/* check for trailing suffix */
-		if(suffix && (n=strlen(suffix)) && n<(last-first))
+		if(suffix && (n=strlen(suffix)) && (ssize_t)n<(last-first))
 		{
 			if(memcmp(last-n,suffix,n)==0)
 				last -=n;
 		}
 	}
 	if(last>first)
-		sfwrite(outfile,first,last-first);
+		sfwrite(outfile,first,(size_t)(last-first));
 	sfputc(outfile,termch);
 }
