+++ src/lib/libcmd/date.c
@@ -2,7 +2,7 @@
 *                                                                      *
 *               This software is part of the ast package               *
 *          Copyright (c) 1992-2012 AT&T Intellectual Property          *
-*          Copyright (c) 2020-2025 Contributors to ksh 93u+m           *
+*          Copyright (c) 2020-2026 Contributors to ksh 93u+m           *
 *                      and is licensed under the                       *
 *                 Eclipse Public License, Version 2.0                  *
 *                                                                      *
@@ -22,6 +22,12 @@
  *
  * date -- set/display date
  */
+
+#include <cmd.h>
+#include <ls.h>
+#include <proc.h>
+#include <tmx.h>
+#include <times.h>
 
 static const char usage[] =
 "[-?\n@(#)$Id: date (AT&T Research) 2011-01-27 $\n]"
@@ -192,12 +198,6 @@ static const char usage[] =
 "	\bstrftime\b(3), \bstrptime\b(3), \btm\b(3)]"
 ;
 
-#include <cmd.h>
-#include <ls.h>
-#include <proc.h>
-#include <tmx.h>
-#include <times.h>
-
 typedef struct Fmt
 {
 	struct Fmt*	next;
@@ -242,7 +242,7 @@ settime(Shbltin_t* context, Time_t now, int adjust, int network)
 	}
 	*argv++ = buf;
 	*argv = 0;
-	if (!sh_run(context, argv - args, args))
+	if (!sh_run(context, (int)(argv - args), args))
 		return 0;
 	return -1;
 }
